Replace placeholders to point to internal services
Use internal service DNS names in config placeholders.

Add a 'local' spring profile to be used when running services outside
the composition network to hit them at localhost instead.

Note the services that export ports to localhost do so only for the sake
of helping in development. In production only the gateway should do
that.