[datviz] Implement index() as a Consumer<GeodataRecord>
Function composition can be applied to generate the input source for the
`index()` consumer, acting as a sink.

For example, given a dataset URI, composing `read` with `toWgs84` and
then sending the stream of records to `index` will create the
appropriate pipeline to use whatever the index consumer implementation
is.

Next step is for the consumer to send the records to ElasticSearch, as
an implementation detail.